+++ packages/api/src/routes/cakes.ts
@@ -2,4 +2,9 @@ import * as express from 'express';
 import { Router } from 'express';
 import { CakesController } from '../controllers/cakes';
 
-export const CakesRoute: Router = express.Router().get('/', CakesController.All);
+const CakesUrl = '/';
+
+export const CakesRoute: Router = express
+  .Router()
+  .get(CakesUrl, CakesController.all)
+  .post(CakesUrl, CakesController.create);