+++ packages/api/src/routes/index.ts
@@ -1,5 +1,6 @@
 import * as express from 'express';
 import { CakesRoute } from './cakes';
+import { RootRoute } from './root';
 
 interface Router {
   path: string;
@@ -10,5 +11,9 @@ export const Router: Router[] = [
   {
     handler: CakesRoute,
     path: '/cakes'
+  },
+  {
+    handler: RootRoute,
+    path: '/'
   }
 ];