+++ packages/api/src/routes/root.ts
@@ -1,0 +1,7 @@
+import * as express from 'express';
+import { Router } from 'express';
+import { CakesController } from '../controllers/cakes';
+
+const Root = '/';
+
+export const RootRoute: Router = express.Router().get(Root, CakesController.allText);