+++ packages/frontend/src/reducers/builder.ts
@@ -1,7 +1,7 @@
 import HandlerBuilder from 'handler-builder';
-import { CakeActions, State, ActionTypes } from '.';
+import { CakesState, UiState, ActionTypes, CakeActions } from './types';
 
-export default (initialState: State) =>
+export default (initialState: CakesState | UiState) =>
   new HandlerBuilder<ActionTypes>(
     (state = initialState, action: CakeActions) => action.type,
     (state = initialState, action: CakeActions) => state