+++ packages/frontend/src/selectors/index.ts
@@ -1,9 +1,8 @@
-import { State } from '../reducers';
+import { createSelector } from 'reselect';
+import { State, UiState } from '../reducers/types';
 
-// I would normally use something like reselect which memoises the result of the function
-// to stop thrashing on mapStateToProps
-// but as the state is so simple there is really no point.
-// All state selections should be done through selectors whether the memoise or not
 export const cakesSelector = (state: State) => state.cakes;
 
-export const isLoadingSelector = (state: State) => state.isLoading;
+export const uiSelector = (state: State) => state.ui;
+
+export const isLoadingSelector = createSelector(uiSelector, (ui: UiState) => ui.isLoading);