+++ packages/frontend/src/services/cakes/index.ts
@@ -11,14 +11,12 @@ export const getCakes = async (): Promise<CakeProps[]> => {
 };
 
 export const persistCake = async (cake: CakeProps) => {
-  const body = JSON.stringify(cake);
-
   await fetch('/cakes', {
     method: 'POST',
     headers: {
       Accept: 'application/json',
       'Content-Type': 'application/json'
     },
-    body
+    body: JSON.stringify(cake)
   });
 };