+++ CHANGELOG.md
@@ -45,6 +45,16 @@
 ## [v0.8.0]
 - [d319441](https://github.com/dagda1/cake-manager/commit/d3194411e6985513f6f0ef30cf118946bfe13637)
 - Hookup cakes form using redux-form, ready to post data to API.
-## [v0.9.9]
+## [v0.9.0]
 - [04f4f9b](https://github.com/dagda1/cake-manager/commit/04f4f9bb2cba84a6b3e1f7a94aed0d340f039c11)
 - install and configure http-proxy-middleware to proxy all non text/html requests to the api
+
+## [v1.0.0]
+- [ed9481c](https://github.com/dagda1/cake-manager/commit/ed9481c39e34dcffbc3083ed871d7a74d8fae2c0)
+-  Return text response from `/` url.  I am presuming what the first requirement is about when it states 
+-  > must be presented in an acceptable format for a human to read
+-  The requirements are ambiguous and open to interpretation but this is my understanding
+-  Add new `/cakes` POST end point.
+-  Hook up the redux-form with the api and call it from the form
+-  Configure proxy to redirect `/` and `Accept: application/json` request to `/cakes` redirects to api
+-  Style cake component